+++ R/modules.R
@@ -29,15 +29,15 @@
 #'       label = "Module",
 #'       module(
 #'         label = "Inner module",
-#'         server = function(id, datasets) {
+#'         server = function(id, data) {
 #'           moduleServer(
 #'             id,
 #'             module = function(input, output, session) {
-#'               output$data <- renderDataTable(datasets$get_data("iris"))
+#'               output$data <- renderDataTable(data[["iris"]]())
 #'             }
 #'           )
 #'         },
-#'         ui = function(id, datasets) {
+#'         ui = function(id) {
 #'           ns <- NS(id)
 #'           tagList(dataTableOutput(ns("data")))
 #'         },
@@ -46,7 +46,7 @@
 #'     ),
 #'     module(
 #'       label = "Another module",
-#'       server = function(id, datasets) {
+#'       server = function(id) {
 #'         moduleServer(
 #'           id,
 #'           module = function(input, output, session) {
@@ -54,7 +54,7 @@
 #'           }
 #'         )
 #'       },
-#'       ui = function(id, datasets) {
+#'       ui = function(id) {
 #'         ns <- NS(id)
 #'         tagList(textOutput(ns("text")))
 #'       },
@@ -156,8 +156,8 @@ is_arg_used.function <- function(modules, arg) {
 #' @param server (`function`) `shiny` module with following arguments:
 #'  - `id` - teal will set proper shiny namespace for this module (see [shiny::moduleServer()]).
 #'  - `input`, `output`, `session` - (not recommended) then [shiny::callModule()] will be used to call a module.
-#'  - `data` (optional) module will receive list of reactive (filtered) data specified in the `filters` argument. `filters`
-#'   can't be NULL
+#'  - `data` (optional) module will receive a `tdata` object, a list of reactive (filtered) data specified in
+#'     the `filters` argument.
 #'  - `datasets` (optional) module will receive `FilteredData`. (See `[teal.slice::FilteredData]`).
 #'  - `reporter` (optional) module will receive `Reporter`. (See [teal.reporter::Reporter]).
 #   - `filter_panel_api` (optional) module will receive `FilterPanelAPI`. (See [teal.slice::FilterPanelAPI]).
@@ -170,8 +170,7 @@ is_arg_used.function <- function(modules, arg) {
 #' @param filters (`character`) A vector with datanames that are relevant for the item. The
 #'   filter panel will automatically update the shown filters to include only
 #'   filters in the listed datasets. `NULL` will hide the filter panel,
-#'   and the keyword `'all'` will show the filters of all datasets. `filters` can't be `NULL` when
-#'   `data` argument is present in the `server` formals.
+#'   and the keyword `'all'` will show the filters of all datasets.
 #' @param server_args (named `list`) with additional arguments passed on to the
 #'   `server` function.
 #' @param ui_args (named `list`) with additional arguments passed on to the
@@ -187,15 +186,15 @@ is_arg_used.function <- function(modules, arg) {
 #'   modules = list(
 #'     module(
 #'       label = "Module",
-#'       server = function(id, datasets) {
+#'       server = function(id, data) {
 #'         moduleServer(
 #'           id,
 #'           module = function(input, output, session) {
-#'             output$data <- renderDataTable(datasets$get_data("iris"))
+#'             output$data <- renderDataTable(data[["iris"]]())
 #'           }
 #'         )
 #'       },
-#'       ui = function(id, datasets) {
+#'       ui = function(id) {
 #'         ns <- NS(id)
 #'         tagList(dataTableOutput(ns("data")))
 #'       }
@@ -207,7 +206,7 @@ is_arg_used.function <- function(modules, arg) {
 #' }
 module <- function(label = "module",
                    server = function(id, ...) {
-                     moduleServer(id, function(input, output, session) {})
+                     moduleServer(id, function(input, output, session) {}) # nolint
                    },
                    ui = function(id, ...) {
                      tags$p(paste0("This module has no UI (id: ", id, " )"))
@@ -246,13 +245,6 @@ module <- function(label = "module",
       "\nFollowing `server_args` elements have no equivalent in the formals of the `server`:\n",
       paste(paste(" -", srv_extra_args), collapse = "\n"),
       "\n\nUpdate the `server` arguments by including above or add `...`"
-    )
-  }
-
-  if ("data" %in% server_formals && is.null(filters)) {
-    stop(
-      "\n`filters = NULL` indicates that the module doesn't need any data while it has the `data` in formals.",
-      "\nPlease specify `filters` with the names of needed datasets or exclude `data` from the arguments."
     )
   }