+++ tests/testthat/test-get_rcode.R
@@ -1,7 +1,8 @@
 testthat::test_that("get_rcode returns header only for empty chunks", {
   ch <- teal.code::chunks_new()
 
-  r_code_from_chunks <- strsplit(get_rcode(chunks = ch), "\n")[[1]]
+  # deprecation warning
+  testthat::expect_warning(r_code_from_chunks <- strsplit(get_rcode(chunks = ch), "\n")[[1]])
   r_code_from_header <- strsplit(sprintf("\n\n%s\n", paste(get_rcode_header(), collapse = "\n")), "\n")[[1]]
 
   # removing the Date line from the header as the seconds may be different
@@ -79,15 +80,16 @@ testthat::test_that("get_datasets_code returns code only for specified datanames
     )
   )
 
+  hashes <- calculate_hashes(datasets$datanames(), datasets)
   testthat::expect_true(
     !grepl(
       "mtcars",
-      paste(get_datasets_code(datasets = datasets, dataname = "IRIS"), collapse = "\n"),
+      paste(get_datasets_code(datasets = datasets, dataname = "IRIS", hashes = hashes), collapse = "\n"),
       ignore.case = TRUE
     ) &&
       grepl(
         "iris",
-        paste(get_datasets_code(datasets = datasets, dataname = "IRIS"), collapse = "\n"),
+        paste(get_datasets_code(datasets = datasets, dataname = "IRIS", hashes = hashes), collapse = "\n"),
         ignore.case = TRUE
       )
   )